Implement searching
Implement searching for definitions using (a) tokens from most fields that have different relevance scores and (b) a trie for quickly finding matching definitions. Multiple search terms are treated as an AND.

Focus the Search box using the '/' key.
Print a definition's category inside the card header in search results (because search results are ordered by relevance, not category)
Ensure that history pushing, and back and forward browser functions works as expected using popstate and hashchange listeners
Hide the intro when a search is being performed

Change link implementation to use template logic rather than a v-html tag
Change URL-management so that all internal state changes are copied /to/ the URL; state is NOT changed by changing the URL
Avoid putting stop words in the search index.
Move the creation of the Vue app into a start() function so it can be run after all other code is declared, but other code can live at the bottom of the file

Add transition group animations to the definitions, which requires instead of /grouping/ definitions by category, /sorting/ them by category, and having the first def in a category gets the category inserted above it.